termcolor: start providing Stderr and StderrLock
To do this without silly code duplication, we add some traits that abstract
over the io::Stdout and io::Stderr types. This ends up being a bit
boiler-platey, but such is life.

After this commit, Stdout becomes a type alias for StandardStream<io::Stdout>,
and StdoutLock becomes a type alias for StandardStreamLock<io::StdoutLock>.

This has only been tested on Linux, so the Windows branches are likely broken
in this commit. To be rectified momentarily.