harmonize ripgrep's -w|--word-regexp with grep's
This fixes https://github.com/BurntSushi/ripgrep/issues/389

grep (and git-grep) allow a match using --word-regexp to begin and end with
non-word characters, by surrounding the match with `(^|[^[:alnum:]_])` and
`([^[:alnum:]_]|$)`.

Before this patch ripgrep, by surrounding the match with `\b`, implicitly
required the match to start with a `\w` character, and would fail to match eg.
the ` - 2` in `1 - 2`, which intuitively should be allowed.

This fixes this behavior by surrounding the match with `(?:^|\W)` and
`(?:\W|$)`, using the Unicode non-word character class `\W` rather than the
ASCII character class `[:alnum:]`