Ignore (one) trailing line break per --file parameter
Currently ripgrep interprets a trailing line break at the end of a
pattern file specified via `--file` or similar as a global "match-all"
pattern. Given that many editors are configured to automatically add a
trailing new line at the end of a file on save, this is not necessarily
a good thing.

This patch removes a single new line from the end of each pattern file
contents by simply dropping an empty rule from the pattern vector if it
is the last rule in the vector after loading rules from a file. Multiple
line breaks are not ignored, and no overhead is added by omitting this
check from within the for loop (not that it would really matter since
this is just init time). Additionally, line breaks not at the end of the
file are similarly kept.