Improve help text for `--sort` and `--sortr` flags
I improved the help documentation in the following manner and for the 
following reasons:
1. It's only logical to put the default sub-option on the first possible 
line, as well as to separately mention that it is indeed the default 
sub-option.
2. Additional options for the flags should describe the main points of 
their purpose without requiring user to read the whole help entry. In my 
opinion, the information sub-options' influence on multi-threading and 
speed are important enough to warrant their inclusion in each 
sub-option's description line text.