These structs are extremely large!
The smallest of them is 600 bytes.
From my (admittedly minimal) knowledge, this causes trashing.

Since these structs are so large, objects before and after this object
won't exist in the same page. Assuming that these values are being used together,
they require 2 pages to actually work instead of (possibly) just 1.

Was the point to prevent a syscall?

Signed-off-by: Hanif Bin Ariffin <hanif.ariffin.4326@gmail.com>